2015-03-15a (A). Relocated resources from buildings.xml to settlements.xml. (B). Enlarged Mars Globe in MainMenu.java.
(1). Changed name of the resource processing "Water Recycling" to "Grey Water Recycling" in buildings.xml.

(2). Corrected the rotation of Mars Globe (now right hand rotation). 

(3). Replaced 8k maps with 2k maps in mars-sim-main module.

+++ mars-sim/mars-sim-ui/src/main/java/org/mars_sim/msp/ui/swing/tool/monitor/VehicleTableModel.java
@@ -13,6 +13,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+
 import javax.swing.SwingUtilities;
 
 import org.mars_sim.msp.core.Coordinates;
@@ -208,7 +209,8 @@ extends UnitTableModel {
 				        String [] availableDesserts = PreparingDessert.getArrayOfDesserts(); 	        
 		    	        for(String n : availableDesserts) {
 		    	        	if (AmountResource.findAmountResource(n).equals(ar)) {
-		    	        		sum += resourceMap.get(ar);
+		    	    			//System.out.println(n + " : "+resourceMap.get(ar));
+		    	        		sum += (double) resourceMap.get(ar);
 		    	        	}
 		    			}
 		    		}
